feat: Support var opts on Waypoint add-ons.
This enables the platform engineer to configure variables on add-on definitions.
