+++ .changelog/170.txt
@@ -1,0 +1,3 @@
+```release-note:feature
+waypoint: Add ability for variable options on an add-on definition to be configured with an HCL file.
+```
