+++ internal/commands/waypoint/add-ons/definitions/create_test.go
@@ -55,6 +55,7 @@ func TestCmdAddOnDefinitionCreate(t *testing.T) {
 				"--readme-markdown-template-file", "readme_test.txt",
 				"--tf-execution-mode", "agent",
 				"--tf-agent-pool-id", "pool-abc123",
+				"--variable-options-file", "vars.hcl",
 			},
 			Expect: &AddOnDefinitionOpts{
 				Name:                        "cli-test",
@@ -67,6 +68,7 @@ func TestCmdAddOnDefinitionCreate(t *testing.T) {
 				ReadmeMarkdownTemplateFile:  "readme_test.txt",
 				TerraformExecutionMode:      "agent",
 				TerraformAgentPoolID:        "pool-abc123",
+				VariableOptionsFile:         "vars.hcl",
 			},
 		},
 	}
@@ -108,6 +110,7 @@ func TestCmdAddOnDefinitionCreate(t *testing.T) {
 				r.Equal(c.Expect.TerraformNoCodeModuleSource, aodOpts.TerraformNoCodeModuleSource)
 				r.Equal(c.Expect.ReadmeMarkdownTemplateFile, aodOpts.ReadmeMarkdownTemplateFile)
 				r.Equal(c.Expect.Labels, aodOpts.Labels)
+				r.Equal(c.Expect.VariableOptionsFile, aodOpts.VariableOptionsFile)
 			}
 		})
 	}
