+++ internal/commands/waypoint/templates/update.go
@@ -9,6 +9,7 @@ import (
 
 	"github.com/hashicorp/hcp-sdk-go/clients/cloud-waypoint-service/preview/2023-08-18/client/waypoint_service"
 	"github.com/hashicorp/hcp-sdk-go/clients/cloud-waypoint-service/preview/2023-08-18/models"
+	"github.com/hashicorp/hcp/internal/commands/waypoint/internal"
 	"github.com/hashicorp/hcp/internal/pkg/cmd"
 	"github.com/hashicorp/hcp/internal/pkg/flagvalue"
 	"github.com/hashicorp/hcp/internal/pkg/heredoc"
@@ -237,9 +238,8 @@ func templateUpdate(opts *TemplateOpts) error {
 
 	// read variable options file and parse hcl
 	var variables []*models.HashicorpCloudWaypointTFModuleVariable
-
 	if opts.VariableOptionsFile != "" {
-		vars, err := parseVariableOptionsFile(opts.VariableOptionsFile)
+		vars, err := internal.ParseVariableOptionsFile(opts.VariableOptionsFile)
 		if err != nil {
 			return errors.Wrapf(err, "%s failed to read Variable Options hcl file %q",
 				opts.IO.ColorScheme().FailureIcon(),
